requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
pydantic>=2.0.0
pyyaml>=6.0
duckdb>=0.9.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
apscheduler>=3.10.0
openpyxl>=3.1.0
altair>=5.0.0